package sercom -> renamed to package ports
